% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr-package.R
\docType{package}
\name{rrr-package}
\alias{rrr}
\alias{rrr-package}
\title{rrr: 'rOpenSci' Review Roclets}
\description{
Companion package to 'rOpenSci' statistical
    software review project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropenscilabs/rrr}
  \item Report bugs at \url{https://github.com/ropenscilabs/rrr/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark@ropensci.org} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

}
\keyword{internal}
