% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist.R
\name{rrr_stats_checklist_check}
\alias{rrr_stats_checklist_check}
\title{rrr_stats_checklist_check}
\usage{
rrr_stats_checklist_check(file)
}
\arguments{
\item{file}{Name of local file containing a completed checklist. Must be a
markdown document in \code{.md} format, not \code{.Rmd} or anything else.}
}
\description{
Correct any potential formatting issues in a completed standards checklist
}
