% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{rrr_stats_roclet}
\alias{rrr_stats_roclet}
\title{rrr_stats_roclet}
\usage{
rrr_stats_roclet()
}
\description{
Get values of all \code{rrrstats} tags in function documentation
}
\details{
Note that this function should never need to be called directly. It only
exists to enable "@rrrstats" tags to be parsed from \pkg{roxygen2}
documentation.
}
