% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rssr-package.R
\docType{package}
\name{rssr-package}
\alias{rssr}
\alias{rssr-package}
\title{rssr: 'rOpenSci' Statistical Software Review}
\description{
Companion package to 'rOpenSci' statistical
    software review project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropenscilabs/rssr}
  \item Report bugs at \url{https://github.com/ropenscilabs/rssr/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark@ropensci.org} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

}
\keyword{internal}
