% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl-standards.R
\name{rssr_available_categories}
\alias{rssr_available_categories}
\title{rssr_available_categories}
\usage{
rssr_available_categories()
}
\value{
A \code{data.frame} with 3 columns of "category" (the categories to be
submitted to \link{rrr_stats_checklist}), "title" (the full title), and
"url".
}
\description{
List all currently available categories and associated URLs to full category
descriptions.
}
