% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-test-package.R
\name{rssr_make_pkg}
\alias{rssr_make_pkg}
\title{rssr_make_pkg}
\usage{
rssr_make_pkg(base_dir = tempdir(), pkg_name = "demo")
}
\arguments{
\item{base_dir}{The base directory where the package should be constructed.}

\item{pkg_name}{The name of the package. The final location of this package
will be in \code{file.path(base_dir, pkg_name)}.}
}
\value{
The path to the directory holding the newly created package
}
\description{
Make a dummy package skeleton including 'rssr' \pkg{roxygen2} tags which can
be used to try out the functionality of this package. Running the example
lines below which activate the 'rssr' roclets, and show you what the output
of those roclets looks like. Feel free to examine the effect of modifying any
of the \verb{@rssr} tags within the code as identified by running those lines.
}
\examples{
d <- rssr_make_pkg ()
roxygen2::roxygenise (d)
}
