% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rssr_roclet.R
\name{rssr_roclet}
\alias{rssr_roclet}
\title{rssr_roclet}
\usage{
rssr_roclet()
}
\description{
Get values of all \code{rssr} tags in function documentation
}
\details{
Note that this function should never need to be called directly. It only
exists to enable "@rssr" tags to be parsed from \pkg{roxygen2} documentation.
}
