% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl-standards.R
\name{rssr_standards_roxygen}
\alias{rssr_standards_roxygen}
\title{rssr_standards_roxygen}
\usage{
rssr_standards_roxygen(
  category = NULL,
  filename = "rssr-standards.R",
  overwrite = FALSE
)
}
\arguments{
\item{category}{One of the names of files given in the directory contents of
\url{https://github.com/ropenscilabs/statistical-software-review-book/tree/master/standards},
each of which is ultimately formatted into a sub-section of the standards.}

\item{filename}{Name of 'R' source file in which to write
\pkg{roxygen2}-formatted lists of standards.}

\item{overwrite}{If \code{FALSE} (default) and \code{filename} already exists, a dialog
will ask whether file should be overwritten.}
}
\value{
Nothing
}
\description{
Obtain rOpenSci standards for statistical software, along with one or more
category-specific standards, as a checklist, convert to project-specific
\pkg{roxygen2} format, and save in nominated file.
}
