% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srr-package.R
\docType{package}
\name{srr-package}
\alias{srr}
\alias{srr-package}
\title{srr: 'rOpenSci' Software Review 'Roclets'}
\description{
Companion package to 'rOpenSci' statistical software review project ('rOpenSci' authors, 2021; \doi{10.5281/zenodo.5556756}), to facilitate documenting software compliance with statistical standards using 'Roxygen2' 'roclets'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/srr/}
  \item \url{https://github.com/ropensci-review-tools/srr}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/srr/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark@ropensci.org} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

Authors:
\itemize{
  \item Mark Padgham \email{mark@ropensci.org} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})
}

}
\keyword{internal}
