% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{srr_release}
\alias{srr_release}
\title{srr_release}
\usage{
srr_release(path)
}
\arguments{
\item{path}{Path to local repository to check}
}
\value{
(Invisibly) List of any standards missing from code
}
\description{
Check that all standards are present in code, and listed either as
'@srrstats' or '@srrstatsNA'
}
