% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{srr_stats_categories}
\alias{srr_stats_categories}
\title{Get details of current statistical software categories}
\usage{
srr_stats_categories()
}
\value{
A \code{data.frame} with 3 columns of "category" (the categories to be
submitted to \link{srr_stats_checklist}), "title" (the full title), and
"url".
}
\description{
List all currently available categories and associated URLs to full category
descriptions.
}
\examples{
srr_stats_categories()
}
\seealso{
Other helper: 
\code{\link{srr_stats_checklist}()},
\code{\link{srr_stats_checklist_check}()},
\code{\link{srr_stats_pkg_skeleton}()},
\code{\link{srr_stats_pre_submit}()}
}
\concept{helper}
