% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl-standards.R
\name{srr_stats_checklist}
\alias{srr_stats_checklist}
\title{Download checklists of statistical software standards}
\usage{
srr_stats_checklist(category = NULL, filename = NULL)
}
\arguments{
\item{category}{One of the names of files given in the directory contents of
\url{https://github.com/ropenscilabs/statistical-software-review-book/tree/main/standards},
each of which is ultimately formatted into a sub-section of the standards.}

\item{filename}{Optional name of local file to save markdown-formatted
checklist. A suffix of \code{.md} will be automatically appended.}
}
\value{
A character vector containing a markdown-style checklist of general
standards along with standards for any additional categories.
}
\description{
Obtain rOpenSci standards for statistical software, along with one or more
category-specific standards, as a checklist, and store the result in the
local clipboard ready to paste.
}
\seealso{
Other helper: 
\code{\link{srr_release}()},
\code{\link{srr_stats_categories}()},
\code{\link{srr_stats_checklist_check}()},
\code{\link{srr_stats_pkg_skeleton}()}
}
\concept{helper}
