% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-submit.R
\name{srr_stats_pre_submit}
\alias{srr_stats_pre_submit}
\title{Perform pre-submission checks}
\usage{
srr_stats_pre_submit(path = ".", quiet = FALSE)
}
\arguments{
\item{path}{Path to local repository to check}

\item{quiet}{If 'FALSE', display information on status of package on screen.}
}
\value{
(Invisibly) List of any standards missing from code
}
\description{
Check that all standards are present in code, and listed either as
'@srrstats' or '@srrstatsNA'
}
\examples{
d <- srr_stats_pkg_skeleton ()
# The skeleton has 'TODO' standards, and also has only a few from the full
# list expected for the categories specified there.
srr_stats_pre_submit (d)
}
\seealso{
Other helper: 
\code{\link{srr_stats_categories}()},
\code{\link{srr_stats_checklist_check}()},
\code{\link{srr_stats_checklist}()},
\code{\link{srr_stats_pkg_skeleton}()}
}
\concept{helper}
