% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{srr_stats_roclet}
\alias{srr_stats_roclet}
\title{srr_stats_roclet}
\usage{
srr_stats_roclet()
}
\description{
Get values of all \code{srrstats} tags in function documentation
}
\details{
Note that this function should never need to be called directly. It only
exists to enable "@srrstats" tags to be parsed from \pkg{roxygen2}
documentation.
}
