% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srr-package.R
\docType{package}
\name{srr-package}
\alias{srr}
\alias{srr-package}
\title{srr: 'rOpenSci' Review Roclets}
\description{
Companion package to 'rOpenSci' statistical software review project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/srr/}
  \item \url{https://github.com/ropensci-review-tools/srr}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/srr/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark@ropensci.org} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

Other contributors:
\itemize{
  \item Mauricio Vargas Sepulveda (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [contributor]
}

}
\keyword{internal}
