% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{srr_release}
\alias{srr_release}
\title{Perform pre-release checks}
\usage{
srr_release(path)
}
\arguments{
\item{path}{Path to local repository to check}
}
\value{
(Invisibly) List of any standards missing from code
}
\description{
Check that all standards are present in code, and listed either as
'@srrstats' or '@srrstatsNA'
}
\seealso{
Other helper: 
\code{\link{srr_stats_categories}()},
\code{\link{srr_stats_checklist_check}()},
\code{\link{srr_stats_checklist}()},
\code{\link{srr_stats_pkg_skeleton}()}
}
\concept{helper}
