% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl-standards.R
\name{srr_stats_categories}
\alias{srr_stats_categories}
\title{srr_stats_categories}
\usage{
srr_stats_categories()
}
\value{
A \code{data.frame} with 3 columns of "category" (the categories to be
submitted to \link{srr_stats_checklist}), "title" (the full title), and
"url".
}
\description{
List all currently available categories and associated URLs to full category
descriptions.
}
