% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist.R
\name{srr_stats_checklist_check}
\alias{srr_stats_checklist_check}
\title{Check a completed standards checklist}
\usage{
srr_stats_checklist_check(file)
}
\arguments{
\item{file}{Name of local file containing a completed checklist. Must be a
markdown document in \code{.md} format, not \code{.Rmd} or anything else.}
}
\description{
Correct any potential formatting issues in a completed standards checklist
}
\seealso{
Other helper: 
\code{\link{srr_stats_categories}()},
\code{\link{srr_stats_checklist}()},
\code{\link{srr_stats_pkg_skeleton}()},
\code{\link{srr_stats_pre_submit}()}
}
\concept{helper}
